Fix SAM3 dtype: use model.half() to match float16 input pipeline
SAM3's io_utils converts images to float16. Model weights default to
bfloat16 on A100. Using .half() aligns both to float16.

+++ pipeline/sam3_segment.py
@@ -234,7 +234,7 @@ class SAM3Backend:
         model = build_sam3_image_model(
             device=builder_device, eval_mode=True,
         )
-        model = model.to(device).float()
+        model = model.to(device).half()
 
         self.processor = Sam3Processor(
             model,
